Removed Fragrance.cs and merged functionality into Molecule
